FINAL VERSION OF LIFEMATE
Some parts are rough around the edges/incomplete.

Login functionality is very basic and does not include a database nor any authentication, at the moment it only simulates what it would be like to login. Also 'logging in' makes the navigation functionality stop working.

Sleep calculator was not fully implemented as we ran out of time to find a way to meaningfully implement it into the program, at the moment we only have a start to its functionality.

Future iterations of the app will have consistent data even when the app is closed and reopened by implemented online database functionality.